WIP actor plan
